Small fixes
[^] board.go | moved UpdateBoards from gochan.go
[^] catalog.go | changed Page type to uint
[^] gochan.go | fixed formatting, moved UpdateBoards to board.go
[^] README.md | added todo list
[^] thread.go | fixed formatting

+++ catalog.go
@@ -12,7 +12,7 @@ type Catalog []CatalogElement
 
 // CatalogElement part of Catalog
 type CatalogElement struct {
-	Page    int64           `json:"page"`
+	Page    uint            `json:"page"`
 	Threads []CatalogThread `json:"threads"`
 }
